Fix testcontainer query with current jaeger
Trying to establish a gRPC connection to 16686 fails with the current jaegertracing/all-in-one:1.24.0 as it is not http2 and presumably not gRPC. As documented in https://www.jaegertracing.io/docs/1.24/deployment/#query-service--ui gRPC query functionality is on port 16685 instead of 16686. This patch adds JAEGER_QUERY_GRPC_PORT and getQueryGrpcPort equivalent to what was there before for the query port and makes use of it in createBlockingQueryService.

Also included is a small robustness improvement that removes the hardcoding of localhost and instead asks the testcontainer for the actual host name in case the mapping is not actually local.